# backend/app/schemas/error.py
from enum import StrEnum
from typing import Literal

from pydantic import BaseModel


class ErrorCode(StrEnum):
    AUTHENTICATION_FAILED = "AUTHENTICATION_FAILED"
    FORBIDDEN = "FORBIDDEN"
    USER_NOT_FOUND = "USER_NOT_FOUND"
    USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS"
    USER_SELF_DEACTIVATION_FORBIDDEN = "USER_SELF_DEACTIVATION_FORBIDDEN"
    LAST_ACTIVE_ADMIN_REQUIRED = "LAST_ACTIVE_ADMIN_REQUIRED"
    PASSWORD_CONFIRMATION_MISMATCH = "PASSWORD_CONFIRMATION_MISMATCH"
    PASSWORD_REUSE_NOT_ALLOWED = "PASSWORD_REUSE_NOT_ALLOWED"
    PASSWORD_RESET_REQUEST_NOT_FOUND = "PASSWORD_RESET_REQUEST_NOT_FOUND"
    PASSWORD_RESET_REQUEST_RESOLVED = "PASSWORD_RESET_REQUEST_RESOLVED"
    PASSWORD_RESET_TARGET_INACTIVE = "PASSWORD_RESET_TARGET_INACTIVE"
    CATEGORY_NOT_FOUND = "CATEGORY_NOT_FOUND"
    CATEGORY_ALREADY_EXISTS = "CATEGORY_ALREADY_EXISTS"
    CATEGORY_INACTIVE = "CATEGORY_INACTIVE"
    TICKET_NOT_FOUND = "TICKET_NOT_FOUND"
    INVALID_TICKET_STATUS_TRANSITION = "INVALID_TICKET_STATUS_TRANSITION"
    INVALID_TICKET_ASSIGNEE = "INVALID_TICKET_ASSIGNEE"
    INVALID_TICKET_FILTER = "INVALID_TICKET_FILTER"
    NOTIFICATION_NOT_FOUND = "NOTIFICATION_NOT_FOUND"
    CUSTOMER_NOT_FOUND = "CUSTOMER_NOT_FOUND"
    CUSTOMER_INACTIVE = "CUSTOMER_INACTIVE"
    CUSTOMER_CONTACT_REQUIRED = "CUSTOMER_CONTACT_REQUIRED"
    CUSTOMER_POTENTIAL_DUPLICATE = "CUSTOMER_POTENTIAL_DUPLICATE"
    CUSTOMER_VERIFICATION_NOT_FOUND = "CUSTOMER_VERIFICATION_NOT_FOUND"
    CUSTOMER_VERIFICATION_INVALID = "CUSTOMER_VERIFICATION_INVALID"
    CUSTOMER_VERIFICATION_FACTOR_UNAVAILABLE = (
        "CUSTOMER_VERIFICATION_FACTOR_UNAVAILABLE"
    )
    VALIDATION_ERROR = "VALIDATION_ERROR"
    SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE"
    INTERNAL_ERROR = "INTERNAL_ERROR"
    BAD_REQUEST = "BAD_REQUEST"
    NOT_FOUND = "NOT_FOUND"
    METHOD_NOT_ALLOWED = "METHOD_NOT_ALLOWED"
    HTTP_ERROR = "HTTP_ERROR"


class ErrorResponse(BaseModel):
    code: ErrorCode
    detail: str


class ValidationErrorItem(BaseModel):
    location: list[str | int]
    message: str
    type: str


class ValidationErrorResponse(BaseModel):
    code: Literal[ErrorCode.VALIDATION_ERROR] = ErrorCode.VALIDATION_ERROR
    detail: str
    errors: list[ValidationErrorItem]