# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthSessionStateResponse": {
        "properties": {
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "session_type",
          "user"
        ],
        "title": "AuthSessionStateResponse",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "AUTHENTICATION_FAILED",
          "FORBIDDEN",
          "USER_NOT_FOUND",
          "USER_ALREADY_EXISTS",
          "USER_SELF_DEACTIVATION_FORBIDDEN",
          "LAST_ACTIVE_ADMIN_REQUIRED",
          "PASSWORD_CONFIRMATION_MISMATCH",
          "PASSWORD_REUSE_NOT_ALLOWED",
          "PASSWORD_RESET_REQUEST_NOT_FOUND",
          "PASSWORD_RESET_REQUEST_RESOLVED",
          "PASSWORD_RESET_TARGET_INACTIVE",
          "CATEGORY_NOT_FOUND",
          "CATEGORY_ALREADY_EXISTS",
          "CATEGORY_INACTIVE",
          "TICKET_NOT_FOUND",
          "INVALID_TICKET_STATUS_TRANSITION",
          "INVALID_TICKET_ASSIGNEE",
          "INVALID_TICKET_FILTER",
          "NOTIFICATION_NOT_FOUND",
          "CUSTOMER_NOT_FOUND",
          "CUSTOMER_INACTIVE",
          "CUSTOMER_CONTACT_REQUIRED",
          "CUSTOMER_POTENTIAL_DUPLICATE",
          "CUSTOMER_VERIFICATION_NOT_FOUND",
          "CUSTOMER_VERIFICATION_INVALID",
          "CUSTOMER_VERIFICATION_FACTOR_UNAVAILABLE",
          "VALIDATION_ERROR",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "BAD_REQUEST",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "HTTP_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NotificationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "NotificationListResponse",
        "type": "object"
      },
      "NotificationReadAllResponse": {
        "properties": {
          "updated_count": {
            "title": "Updated Count",
            "type": "integer"
          }
        },
        "required": [
          "updated_count"
        ],
        "title": "NotificationReadAllResponse",
        "type": "object"
      },
      "NotificationResponse": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_read": {
            "title": "Is Read",
            "type": "boolean"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "ticket": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationTicketReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          }
        },
        "required": [
          "id",
          "type",
          "is_read",
          "created_at",
          "read_at",
          "actor",
          "ticket"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
      "NotificationTicketReference": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ticket_number": {
            "title": "Ticket Number",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticket_number",
          "title"
        ],
        "title": "NotificationTicketReference",
        "type": "object"
      },
      "NotificationType": {
        "enum": [
          "TICKET_ASSIGNED",
          "TICKET_REASSIGNED",
          "TICKET_PUBLIC_COMMENT",
          "TICKET_IN_PROGRESS",
          "TICKET_RESOLVED",
          "TICKET_CLOSED",
          "PASSWORD_RESET_REQUESTED",
          "PASSWORD_RESET_COMPLETED"
        ],
        "title": "NotificationType",
        "type": "string"
      },
      "NotificationUnreadCountResponse": {
        "properties": {
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "unread_count"
        ],
        "title": "NotificationUnreadCountResponse",
        "type": "object"
      },
...
      "SessionType": {
        "enum": [
          "NORMAL",
          "PASSWORD_CHANGE_REQUIRED"
        ],
        "title": "SessionType",
        "type": "string"
      },
...
      "UserReference": {
        "properties": {
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name"
        ],
        "title": "UserReference",
        "type": "object"
      },
      "UserResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "email",
          "first_name",
          "last_name",
          "role",
          "id",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "EMPLOYEE",
          "AGENT",
          "ADMIN"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationErrorItem": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ValidationErrorItem",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "const": "VALIDATION_ERROR",
            "default": "VALIDATION_ERROR",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "detail",
          "errors"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PulseDesk API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A temporary service failure occurred."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Current User",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "list_notifications",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A temporary service failure occurred."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "operationId": "mark_all_notifications_read",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationReadAllResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A temporary service failure occurred."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark All Notifications Read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/unread-count": {
      "get": {
        "operationId": "get_notification_unread_count",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationUnreadCountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A temporary service failure occurred."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Notification Unread Count",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "patch": {
        "operationId": "mark_notification_read",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Notification Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication is required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Owned notification not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "An unexpected server error occurred."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "A temporary service failure occurred."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Notification Read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  }
}